+++ chat-widget/src/components/startchaterrorpanestateful/common/defaultStartChatErrorPaneIconImageProps.ts
@@ -2,7 +2,7 @@ import { IImageProps } from "@fluentui/react";
 import { ImageFit } from "@fluentui/react";
 import { AlertIcon } from "../../../assets/Icons";
 
-export const defaultErrorPaneIconImageStyleProps: IImageProps = {
+export const defaultStartChatErrorPaneIconImageStyleProps: IImageProps = {
     src: AlertIcon,
     imageFit: ImageFit.centerContain,
     width: "86px",