+++ chat-widget/src/components/startchaterrorpanestateful/common/defaultStartChatErrorPaneTitleStyleProps.ts
@@ -1,12 +1,12 @@
 import { IStyle } from "@fluentui/react";
 
-export const defaultErrorPaneSubtitleStyleProps: IStyle = {
+export const defaultStartChatErrorPaneTitleStyleProps: IStyle = {
     fontFamily: "'Segoe UI',Arial,sans-serif",
     fontWeight: "normal",
     fontSize: "18px",
     color: "#36454F",
     textAlign: "center",
     display: "flex",
-    order: 3,
+    order: 2,
     alignSelf: "auto"
 };