+++ CHANGE_LOG.md
@@ -233,7 +233,7 @@ All notable changes to this project will be documented in this file.
 
 ### Added
 
-- Add ability to customize `PreChatSurveyPane`'s toggle inputs using `IPreChatSurveyPaneToggleInputStyles`
+- Add ability to customize `PreChatSurveyPane`'s toggle inputs using `IPreChatSurveyPaneToggleInputStyles` & Use `baseline` by default
 
 ### Fixed