+++ CHANGE_LOG.md
@@ -7,7 +7,7 @@ All notable changes to this project will be documented in this file.
 ## [Unreleased]
 
 ### Added
-- Enabled customizations for the start chat error pane
+- Enabled customizations for the start chat error pane by adding new `IStartChatErrorPaneProps` interface
 
 ### Fixed